Add SetCharacterSize function
This implements setting the character width and height via the
`GS ! n` command.  The valid values for both width and height are from 0
to 7, inclusively.

While the printer accepts all these values, it may not print them
correctly.  On the printer that I have all widths work but any height 4
and above will not print correctly.