[printhis] Add ESCPOS_DEVICE environment variable
The printhis utility will now look at the ESCPOS_DEVICE environment
variable for the device path or IP address.  The value must be either
tcp://[ip or host]:[port] or file://[path to device file].

If both the environment variable and --addr or --device are specified,
the command line flags take precedent.